// fhf-datafangst-client/src/generated/openapi/models/fuel-measurement-or-bunkering-one-of.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { FuelMeasurementOrBunkeringOneOfValue } from './fuel-measurement-or-bunkering-one-of-value';

/**
 * 
 * @export
 * @interface FuelMeasurementOrBunkeringOneOf
 */
export interface FuelMeasurementOrBunkeringOneOf {
    /**
     * 
     * @type {string}
     * @memberof FuelMeasurementOrBunkeringOneOf
     */
    'type': FuelMeasurementOrBunkeringOneOfTypeEnum;
    /**
     * 
     * @type {FuelMeasurementOrBunkeringOneOfValue}
     * @memberof FuelMeasurementOrBunkeringOneOf
     */
    'value': FuelMeasurementOrBunkeringOneOfValue;
}

export const FuelMeasurementOrBunkeringOneOfTypeEnum = {
    Bunkering: 'bunkering'
} as const;

export type FuelMeasurementOrBunkeringOneOfTypeEnum = typeof FuelMeasurementOrBunkeringOneOfTypeEnum[keyof typeof FuelMeasurementOrBunkeringOneOfTypeEnum];