// fhf-datafangst-client/src/generated/openapi/models/fuel-measurement-or-bunkering-one-of1.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { FuelMeasurementOrBunkeringOneOfValue } from './fuel-measurement-or-bunkering-one-of-value';

/**
 * 
 * @export
 * @interface FuelMeasurementOrBunkeringOneOf1
 */
export interface FuelMeasurementOrBunkeringOneOf1 {
    /**
     * 
     * @type {string}
     * @memberof FuelMeasurementOrBunkeringOneOf1
     */
    'type': FuelMeasurementOrBunkeringOneOf1TypeEnum;
    /**
     * 
     * @type {FuelMeasurementOrBunkeringOneOfValue}
     * @memberof FuelMeasurementOrBunkeringOneOf1
     */
    'value': FuelMeasurementOrBunkeringOneOfValue;
}

export const FuelMeasurementOrBunkeringOneOf1TypeEnum = {
    FuelMeasurement: 'fuelMeasurement'
} as const;

export type FuelMeasurementOrBunkeringOneOf1TypeEnum = typeof FuelMeasurementOrBunkeringOneOf1TypeEnum[keyof typeof FuelMeasurementOrBunkeringOneOf1TypeEnum];